Align httpd_query_key_value behavior wrt. string length.
The `val_size` is not a pointer to `size_t`, and if the input
buffer size is too small, will put the needed size in it.
The behavior is now aligned with `httpd_cookie_key_value`.

Also adds extra checks and allow input buffer to be a NULL pointer,
for cases where the caller is only interested in getting the size.